+++ lib/assets.php
@@ -49,6 +49,11 @@ add_action(
 		/* Theme styles and scripts */
 		$space_assets->enqueue( 'theme', 'main', [] );
 
+		/* Floating header scripts */
+		if ( 'floating-header' === get_theme_mod( 'space_header_options', false ) ) {
+			$space_assets->enqueue( 'theme', 'floatingHeader', [] );
+		}
+
 		/* Comment reply js */
 		if (
 		is_single() &&